requirements.txt
astroid==2.1.0
atomicwrites==1.3.0
attrs==18.2.0
coverage==4.5.2
Django==2.1.7
django-cors-headers==2.4.0
django-environ==0.4.5
django-extensions==2.1.5
djangorestframework==3.9.1
factory-boy==2.11.1
Faker==1.0.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==6.0.0
pluggy==0.8.1
psycopg2==2.7.7
py==1.8.0
pylint==2.2.2
pytest==4.3.1
pytest-cov==2.6.1
pytest-django==3.4.7
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
text-unidecode==1.2
typed-ast==1.3.1
wrapt==1.11.1
